requirements.txt
# Core Flask dependencies
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Vector database and AI dependencies
chromadb>=0.4.15
openai>=1.12.0
sentence-transformers>=2.2.2
numpy>=1.24.3
pandas>=2.0.3
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.2
faiss-cpu>=1.7.4

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Security dependencies
PyJWT>=2.8.0
cryptography>=41.0.0
bleach>=6.0.0

# File validation dependencies (optional but recommended)
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == "Windows"

# Rate limiting (optional)
flask-limiter>=3.5.0
redis>=4.5.0

# Form validation (optional)
wtforms>=3.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0
